requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
pdfplumber>=0.10.3
numpy>=1.26.2
anthropic>=0.7.8
faiss-cpu>=1.13.2
sentence-transformers>=2.2.2
tiktoken>=0.5.2
